help windows

+++ controller/CardController.py
@@ -8,7 +8,8 @@ from PIL import Image, ImageTk
 from tkinter import filedialog as fd
 from num2words import num2words
 import playsound        # pip install --upgrade setuptools wheel  # pip install playsound
-from gtts import gTTS   # pip install gTTS
+from gtts import gTTS
+import requests   # pip install gTTS
 
 from controller.ProductController import ProductController
 from model.Category import Category
@@ -128,9 +129,9 @@ class CardController:
                 self.paid.set(self.cardTotal.get())
                 self.profit.set(Card.profit(self))
                 self.bill_box.delete('1.0', tk.END)
-                self.bill_box.insert(tk.END, '\t\t\t'+str(shop[0])+'\n')
-                self.bill_box.insert(tk.END, '\t\t\t        '+str(shop[1])+'\n')
-                self.bill_box.insert(tk.END, '\t\t\t           Mobile : - '+str(shop[2])+'\n')
+                self.bill_box.insert(tk.END, str(shop[0])+'\n',"center")
+                self.bill_box.insert(tk.END, str(shop[1])+'\n', "center")
+                self.bill_box.insert(tk.END, 'Mobile : - '+str(shop[2])+'\n', "center")
                 messagebox.showinfo("Success", " Successfully refresh your card")
 
 
@@ -163,29 +164,45 @@ class CardController:
             if list:
 
                 self.bill_box.delete('1.0', tk.END)
-                self.bill_box.insert(tk.END, '\t\t\t'+str(shop[0])+'\n')
-                self.bill_box.insert(tk.END, '\t\t\t        '+str(shop[1])+'\n')
-                self.bill_box.insert(tk.END, '\t\t\t           Mobile : - '+str(shop[2])+'\n')
+                self.bill_box.insert(tk.END, str(shop[0])+'\n',"center")
+                self.bill_box.insert(tk.END, str(shop[1])+'\n', "center")
+                self.bill_box.insert(tk.END, 'Mobile : - '+str(shop[2])+'\n', "center")
 
-                self.bill_box.insert(tk.END, '\n---------------------------------------------  SHOPPING CARD :    -----------------------------------------')
+                self.bill_box.insert(tk.END, '\n---------------------------------------------  SHOPPING CARD :    --------------------------------')
                 self.bill_box.insert(tk.END, '\nCustomer Name :  '+str(self.name.get())+'')
                 self.bill_box.insert(tk.END, '\nMobile Number :    '+str(self.mobile.get())+'')
-                self.bill_box.insert(tk.END, "\n--------------------------------------------------------------------------------------------------------------------\n")
+                self.bill_box.insert(tk.END, "\n-----------------------------------------------------------------------------------------------------------\n")
                 self.bill_box.insert (tk.END, "DESCRIPTION  \t\t\t RATE \t QUANTITY \t\t AMOUNT")
-                self.bill_box.insert(tk.END, "\n--------------------------------------------------------------------------------------------------------------------\n")
+                self.bill_box.insert(tk.END, "\n-----------------------------------------------------------------------------------------------------------\n")
                 for lists in list:
                     self.bill_box.insert (tk.END, ''+str(lists[1])+'  \t\t\t '+str(lists[2])+' \t '+str(lists[3])+' \t\t '+str(lists[2]*lists[3])+' \n' )
-                self.bill_box.insert(tk.END, "\n--------------------------------------------------------------------------------------------------------------------\n")
+                self.bill_box.insert(tk.END, "\n-----------------------------------------------------------------------------------------------------------\n")
                 self.bill_box.insert (tk.END, '\t\t\t\tTOTAL PRICE \t  = \t '+str(self.cardTotal.get())+'\n')
                 self.bill_box.insert (tk.END, '\t\t\t\tLess  (-)\t            = \t '+str(self.less.get())+'\n ')
-                self.bill_box.insert (tk.END,"\t\t------------------------------------------------------------------------------------\n" )
+                self.bill_box.insert (tk.END,"\t\t-------------------------------------------------------------------------------\n" )
                 self.bill_box.insert (tk.END, '\t\t\t\tNET PAYABLE \t  = \t '+str(self.cardTotal.get()-self.less.get())+'\n' )
                 self.bill_box.insert (tk.END, '\t\t\t\tPaid   (-)\t           = \t '+str(self.cardTotal.get()-self.less.get()-self.due.get())+'\n' )
-                self.bill_box.insert (tk.END, "\t\t      ------------------------------------------------------------------------\n" )
+                self.bill_box.insert (tk.END, "\t\t      ---------------------------------------------------------------------------\n" )
                 self.bill_box.insert (tk.END, '\t\t\t\tDUE AMOUNT   \t=\t '+str(self.due.get())+'\n ')
-
+    def checkConnection(self):
+        url = "http://www.google.com"
+        timeout = 50
+        try:
+            response = requests.get(url, timeout=timeout)
+            if response.status_code == 200:
+                return True
+            else:
+                return False
+        except requests.ConnectionError:
+            return False
+        except requests.Timeout:
+            return False
 
     def cardSpeaker(self):
+        ctotal = self.cardTotal.get()
+        less   = self.less.get()
+        paid   = self.paid.get()
+        due    = self.due.get()
         if self.volume_image["toggle"]:
             self.volume_button.config(image=self.au_of)
             self.volume_button.image = self.au_of
@@ -193,13 +210,23 @@ class CardController:
             list = Card.all(self)
             texts = ''
             if list:
-                for lists in list:
-                        texts = '   '+str(lists[1])+'  price   '+num2words(str(lists[2]), lang='en-US')+'    quantity   '+num2words(str(lists[3]), lang='en-US')+' \n\n\n'
+                if CardController.checkConnection(self):
+                    for lists in list:
+                        texts = '. Product name.'+str(lists[1])+'.  price .  '+num2words(str(lists[2]), lang='en-US')+'.    quantity.   '+num2words(str(lists[3]), lang='en-US')+'. Total Amount.  '+num2words(str(lists[2]*lists[3]), lang='en-US')+'. \n'
                         toSpeak = gTTS(text=texts, lang='en-US', tld='us')
                         file = os.path.abspath(os.path.expanduser( '~' )+"\\AppData\\Local\\BMS\\audio.mp3")
                         toSpeak.save(file)
                         playsound.playsound(file, True)
                         os.remove(file)
+                        
+                    texts = 'Total Price.'+num2words(str(ctotal), lang='en-US')+'.  Total Discount .  '+num2words(str(less), lang='en-US')+'.    Total Payable Amount.   '+num2words(str(paid), lang='en-US')+'. Total Due .  '+num2words(str(due), lang='en-US')+'.  Thanks You.'
+                    toSpeak = gTTS(text=texts, lang='en-US', tld='us')
+                    file = os.path.abspath(os.path.expanduser( '~' )+"\\AppData\\Local\\BMS\\audio.mp3")
+                    toSpeak.save(file)
+                    playsound.playsound(file, True)
+                    os.remove(file)
+                else:
+                    messagebox.showwarning("Info","Internet connection required!")
         else:
             self.volume_button.config(image=self.au_on)
             self.volume_button.image = self.au_on
